+++ ydb/core/persqueue/pqtablet/pq_impl.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "percentile_counter.h"
-#include "metering_sink.h"
-#include "transaction.h"
+#include <ydb/core/persqueue/percentile_counter.h>
+#include <ydb/core/persqueue/metering_sink.h>
+#include <ydb/core/persqueue/transaction.h>
 
 #include <ydb/core/keyvalue/keyvalue_flat_impl.h>
 #include <ydb/core/tablet/tablet_counters.h>
