+++ ydb/core/persqueue/pqtablet/pq_impl_app.cpp
@@ -1,8 +1,8 @@
 #include "pq_impl.h"
 #include "pq_impl_types.h"
 
-#include "common_app.h"
-#include "partition_log.h"
+#include <ydb/core/persqueue/common_app.h>
+#include <ydb/core/persqueue/partition_log.h>
 
 #include <ydb/library/protobuf_printer/security_printer.h>
